# logisland-framework/logisland-agent/src/main/swagger/api-swagger.yaml
swagger: '2.0'
info:
  description: REST API for logisland agent
  version: v1
  title: logisland-agent
  contact:
    name: Thomas Bailet
    email: bailet.thomas@gmail.com
host: localhost:8081
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /topics:
    get:
      tags:
        - topic
      summary: get all topics
      operationId: getAllTopics
      parameters: []
      responses:
        '200':
          description: Status 200
          schema:
            type: array
            items:
              $ref: '#/definitions/Topic'
    post:
      tags:
          - topic
      summary: create new topic
      operationId: addNewTopic
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Topic'
      responses:
        '200':
          description: Status 200
  /topics/{topicId}:
    get:
      tags:
        - topic
      summary: get topic
      operationId: getTopic
      parameters:
        - name: topicId
          in: path
          required: true
          type: string
      responses:
        "200":
          description: Status 200
          schema:
            $ref: '#/definitions/Topic'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
          - topic
      summary: update topic
      operationId: updateTopic
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Topic'
        - name: topicId
          in: path
          required: true
          type: string
      responses:
        "200":
          description: job successfuly started
          schema:
            $ref: '#/definitions/Topic'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - topic
      summary: delete topic
      description: remove a topic config and remove all content from Kafka
      operationId: deleteTopic
      parameters:
        - name: topicId
          in: path
          required: true
          type: string
      responses:
        "200":
          description: topic successfully deleted
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /topics/{topicId}/keySchema:
    get:
      tags:
        - schema
      summary: get topic key schema
      operationId: getTopicKeySchema
      parameters:
        - name: topicId
          in: path
          required: true
          type: string
        - name: version
          in: query
          description: version of the schema ("latest" if not provided)
          required: false
          type: string
          default: latest
      responses:
        "200":
          description: Avro schema
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - schema
      summary: update topic key schema
      operationId: updateTopicKeySchema
      parameters:
        - name: body
          in: body
          description: schema to add to the store
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: Avro schema
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /topics/{topicId}/keySchema/checkCompatibility:
    post:
      tags:
        - schema
      summary: check topic key schema compatibility
      operationId: checkTopicKeySchemaCompatibility
      parameters:
        - name: body
          in: body
          description: Avro schema as a json string
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: compatibility level
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /topics/{topicId}/valueSchema:
    get:
      tags:
        - schema
      summary: get topic value schema
      operationId: getTopicValueSchema
      parameters:
        - name: topicId
          in: path
          description: id of the job to return
          required: true
          type: string
        - name: version
          in: query
          description: version of the schema ("latest" if not provided)
          required: false
          type: string
          default: latest
      responses:
        "200":
          description: job definition
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - schema
      summary: update topic value schema
      operationId: updateTopicValueSchema
      parameters:
        - name: body
          in: body
          description: Avro schema as a json string
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          description: id of the job to return
          required: true
          type: string

      responses:
        "200":
          description: Avro schema
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /topics/{topicId}/valueSchema/checkCompatibility:
    post:
      tags:
        - schema
      summary: check topic value schema compatibility
      operationId: checkTopicValueSchemaCompatibility
      parameters:
        - name: topicId
          in: path
          description: id of the job to return
          required: true
          type: string
        - name: body
          in: body
          description: Avro schema as a json string
          required: true
          schema:
            type: string
      responses:
        "200":
          description: compatibility level
          schema:
            type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  Field:
    type: object
    required:
      - name
      - type
    properties:
      name:
        description: a unique identifier for the topic
        type: string
      encrypted:
        description: is the field need to be encrypted
        type: boolean
      indexed:
        description: is the field need to be indexed to search store
        type: boolean
      persistent:
        description: is the field need to be persisted to data store
        type: boolean
      type:
        description: the type of the field
        type: string
        default: string
        enum:
        - null
        - string
        - long
        - array
        - float
        - double
        - bytes
        - record
        - map
        - enum
        - boolean
      value:
        description: a unique identifier for the topic
        type: object
  Record:
    type: object
    required:
      - name
    properties:
      id:
        description: a unique identifier for the topic
        type: integer
        format: int64
      name:
        description: the name of the topic
        type: string
      multiline:
        description: is the record multiline
        type: boolean
      timestampField:
        description: the record_time field
        type: string
        default: "record_time"
      rowkeyField:
        description: the record_id field
        type: string
        default: "record_id"
      fields:
        type: array
        items:
          $ref: '#/definitions/Field'
...
  Topic:
    type: object
    required:
      - name
      - partitions
      - replicationFactor
      - valueSchema
      - serializer
    properties:
      id:
        description: a unique identifier for the topic
        type: integer
        format: int64
      version:
        description: the version of the topic configuration
        type: integer
        format: int32
      name:
        description: the name of the topic
        type: string
      serializer:
        description: the class of the Serializer
        type: string
      valueSchema:
        $ref: '#/definitions/Record'
      keySchema:
        $ref: '#/definitions/Record'
      partitions:
        description: default number of partitions
        type: integer
        format: int64
      replicationFactor:
        description: default replication factor
        type: integer
        format: int64
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string